Bro-Hiking/mqtt_app/requirements.txt
paho-mqtt
urllib3
pymongo